+++ src/mui/field/index.js
@@ -2,6 +2,8 @@ export BooleanField from './BooleanField';
 export ChipField from './ChipField';
 export DateField from './DateField';
 export EmailField from './EmailField';
+export FunctionField from './FunctionField';
+export NumberField from './NumberField';
 export ReferenceField from './ReferenceField';
 export ReferenceManyField from './ReferenceManyField';
 export RichTextField from './RichTextField';