+++ src/mui/input/BooleanInput.js
@@ -1,5 +1,6 @@
 import React, { PropTypes } from 'react';
 import Toggle from 'material-ui/Toggle';
+import title from '../../util/title';
 
 const styles = {
     block: {
@@ -14,22 +15,24 @@ const styles = {
     },
 };
 
-const BooleanInput = ({ input, label }) => (
-    <div style={styles.block}>
+const BooleanInput = ({ input, label, source, elStyle }) => (
+    <div style={elStyle || styles.block}>
         <Toggle
             defaultToggled={!!input.value}
             onToggle={input.onChange}
             labelStyle={styles.label}
             style={styles.toggle}
-            label={label}
+            label={title(label, source)}
         />
     </div>
 );
 
 BooleanInput.propTypes = {
+    elStyle: PropTypes.object,
     includesLabel: PropTypes.bool.isRequired,
     input: PropTypes.object,
     label: PropTypes.string,
+    source: PropTypes.string.isRequired,
 };
 
 BooleanInput.defaultProps = {