+++ src/mui/input/NullableBooleanInput.spec.js
@@ -11,7 +11,7 @@ describe('<NullableBooleanInput />', () => {
     };
 
     it('should give three different choices for true, false or unknown', () => {
-        const wrapper = shallow(<NullableBooleanInput {...defaultProps} />);
+        const wrapper = shallow(<NullableBooleanInput source="foo" {...defaultProps} />);
         const choices = wrapper.find('SelectInput').prop('choices');
         assert.deepEqual(choices, [
             { id: null, name: '' },
@@ -22,19 +22,19 @@ describe('<NullableBooleanInput />', () => {
 
     describe('error message', () => {
         it('should not be displayed if field is pristine', () => {
-            const wrapper = shallow(<NullableBooleanInput {...defaultProps} meta={{ touched: false }} />);
+            const wrapper = shallow(<NullableBooleanInput source="foo" {...defaultProps} meta={{ touched: false }} />);
             const SelectInputElement = wrapper.find('SelectInput');
             assert.equal(SelectInputElement.prop('errorText'), false);
         });
 
         it('should not be displayed if field has been touched but is valid', () => {
-            const wrapper = shallow(<NullableBooleanInput {...defaultProps} meta={{ touched: true, error: false }} />);
+            const wrapper = shallow(<NullableBooleanInput source="foo" {...defaultProps} meta={{ touched: true, error: false }} />);
             const SelectInputElement = wrapper.find('SelectInput');
             assert.equal(SelectInputElement.prop('errorText'), false);
         });
 
         it('should be displayed if field has been touched and is invalid', () => {
-            const wrapper = shallow(<NullableBooleanInput {...defaultProps} meta={{ touched: true, error: 'Required field.' }} />);
+            const wrapper = shallow(<NullableBooleanInput source="foo" {...defaultProps} meta={{ touched: true, error: 'Required field.' }} />);
             const SelectInputElement = wrapper.find('SelectInput');
             assert.equal(SelectInputElement.prop('errorText'), 'Required field.');
         });